pyproject.toml
[project]
name = "agente-presupuesto-telegram"
version = "0.1.0"
description = "Bot de Telegram para presupuestos de obra con MiniMax-M2 como orquestador"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot[webhooks]>=21.6",
    "pydantic>=2.8",
    "pydantic-settings>=2.4",
    "python-dotenv>=1.0",
    "pandas>=2.2",
    "jinja2>=3.1",
    "weasyprint>=62.3",
    "pyyaml>=6.0",
    "openai>=1.52",
    "httpx>=0.27",
    "structlog>=24.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "hypothesis>=6.112",
    "ruff>=0.6",
    "mypy>=1.11",
    "types-PyYAML",
    "pandas-stubs",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-ra --strict-markers"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
